SAS Stabilization of Undesired Torque
Implemented a system that nullifies and/or dampens the effects of torque on control surfaces given certain airspeed values (nullifies below 9 m/s, dampens below 27 m/s).

This makes it possible to make a stable liftoff, a stable landing, and retain some control against wind shears (which will be an apparently common issue).